# cache: Not sending info request for cached chunk