Improve cache backend performance.