@ncw 2018-08-28
I'm not sure this is worth doing at most it will save a few cycles and stop contention on the refreshMutex.  Have you measured an improvement with this patch?