@dcpu 2018-08-28
rclone always sends an info request to the source remote, regardless of whether the chunk is already cached.
Therefore, access to the cached chunk is delayed by latency between source remote.
This patch prevents this. The difference is noticable when you access already cached chunks.