@dcpu 2018-08-29
I wrote a test program for random read performance benchmarks.
Mount cache remote and run 

randtest [file to cache mount]

```
package main

import (
	"log"
	"math/rand"
	"os"
	"time"
)

func check(e error) {
	if e != nil {
		panic(e)
	}
}

func main() {
	var err error
	b := make([]byte, 16384)
	fn := os.Args[1]
	f, err := os.Open(fn)
	check(err)
	fi, err := f.Stat()
	check(err)
	log.Printf(fn)

	start := time.Now()
	fs := fi.Size()
	for i := 0; i < 30; i++ {
		cstart := time.Now()
		of := rand.Int63n(fs)
		_, err = f.Seek(of, 0)
		check(err)
		_, err = f.Read(b)
		check(err)
		celapsed := time.Since(cstart)
		log.Printf("offset %d took %s", of, celapsed)
	}
	f.Close()
	elapsed := time.Since(start)
	log.Printf("took %s", elapsed)
}
```

Here is my results from 140M file on gdrivecache mount.

Original

1m 8s non cached
42s cached

With #2493

42s non cached
12s cached

With #2493 #2495

32s non cached
0.6s cached