@ncw 2018-08-30
Thanks for doing the test :-)

I was looking at the source of `refreshFromSource` - I can see I mis-read it now and your change does make a difference.

So it looks like your change is equivalent to calling `o.refreshFromSource(false)` rather than `o.refreshFromSource(true)` - would you agree?

```go
// refreshFromSource requests the original FS for the object in case it comes from a cached entry
func (o *Object) refreshFromSource(force bool) error {
	o.refreshMutex.Lock()
	defer o.refreshMutex.Unlock()
	var err error
	var liveObject fs.Object

	if o.Object != nil && !force {
		return nil
	}
```

@remusb - can you help with this?  Is this sensible or will we miss updates from the source?