@dcpu 2018-08-30
Yes. They are equivalent.