@B4dM4n 2018-09-03
This looks good to me.
Using `o.refresh()` is the right way to check for expiration when `o.Object != nil`.

What I didn't figure out yet is why this change impacts seek performance. `Object.Open` should only be called once, but it looks like it is called on every seek.

I will investigate this further.