@B4dM4n 2018-09-03
Got it, `cache.Object.Open` returns a `readers.NewLimitedReadCloser` not a `cache.Object.Handle` as I assumed.

This change definitely improves the cache performance.