@ncw 2018-09-03
Thanks for your review @B4dM4n 

I'll merge this now then - thank you very much @dcpu for the contribution.